requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.9.1
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
discord.py==2.0.1
idna==3.3
multidict==6.0.2
python-dotenv==1.0.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
yarl==1.7.2
